% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate Directories.R
\name{generate_directory}
\alias{generate_directory}
\title{Create R project directory structure}
\usage{
generate_directory(loc = NULL)
}
\arguments{
\item{loc}{Location to create the structure in. If \code{NULL}, automatically assumes
current working directory}
}
\description{
Creates a directory structure for data analysis projects
}

